+++ RX/MDK-ARM/RX/RX.build_log.htm
@@ -21,13 +21,13 @@ Target DLL:      STLink\ST-LINKIII-KEIL_SWO.dll V3.0.7.0
 Dialog DLL:      TCM.DLL V1.46.0.0
  
 <h2>Project:</h2>
-D:\khoa\3\RX\MDK-ARM\RX.uvprojx
-Project File Date:  06/02/2023
+D:\khoa\6\RX\MDK-ARM\RX.uvprojx
+Project File Date:  06/03/2023
 
 <h2>Output:</h2>
 *** Using Compiler 'V5.06 update 6 (build 750)', folder: 'D:\phanmem\Keil_v5\ARM\ARMCC\Bin'
 Build target 'RX'
-compiling main.c...
+compiling NRF24L01.c...
 linking...
 Program Size: Code=5188 RO-data=304 RW-data=36 ZI-data=1716  
 FromELF: creating hex file...
@@ -55,7 +55,7 @@ Package Vendor: Keil
 
    * Component: ARM::CMSIS:CORE:5.4.0
       Include file:  CMSIS\Core\Include\tz_context.h
-Build Time Elapsed:  00:00:02
+Build Time Elapsed:  00:00:01
 </pre>
 </body>
 </html>
