+++ TX/MDK-ARM/TX/TX.build_log.htm
@@ -21,17 +21,12 @@ Target DLL:      STLink\ST-LINKIII-KEIL_SWO.dll V3.0.7.0
 Dialog DLL:      TCM.DLL V1.46.0.0
  
 <h2>Project:</h2>
-D:\khoa\6\TX\MDK-ARM\TX.uvprojx
+D:\khoa\5\TX\MDK-ARM\TX.uvprojx
 Project File Date:  06/03/2023
 
 <h2>Output:</h2>
 *** Using Compiler 'V5.06 update 6 (build 750)', folder: 'D:\phanmem\Keil_v5\ARM\ARMCC\Bin'
 Build target 'TX'
-compiling NRF24L01.c...
-compiling main.c...
-linking...
-Program Size: Code=5480 RO-data=328 RW-data=32 ZI-data=1744  
-FromELF: creating hex file...
 "TX\TX.axf" - 0 Error(s), 0 Warning(s).
 
 <h2>Software Packages used:</h2>
@@ -55,7 +50,7 @@ Package Vendor: Keil
 <h2>Collection of Component Files used:</h2>
 
    * Component: ARM::CMSIS:CORE:5.4.0
-Build Time Elapsed:  00:00:01
+Build Time Elapsed:  00:00:00
 </pre>
 </body>
 </html>
